src/ml_core/evaluate.py
"""Evaluation metrics for institutional ML binary classification."""

from __future__ import annotations

from typing import Any

import numpy as np
from sklearn.metrics import (
    accuracy_score,
    confusion_matrix,
    f1_score,
    precision_score,
    recall_score,
    roc_auc_score,
)


def compute_metrics(y_true, y_pred, y_proba) -> dict[str, float]:
    """Compute binary classification metrics."""

    metrics = {
        "accuracy": float(accuracy_score(y_true, y_pred)),
        "precision": float(precision_score(y_true, y_pred, zero_division=0)),
        "recall": float(recall_score(y_true, y_pred, zero_division=0)),
        "f1": float(f1_score(y_true, y_pred, zero_division=0)),
    }
    try:
        metrics["roc_auc"] = float(roc_auc_score(y_true, y_proba))
    except Exception:
        metrics["roc_auc"] = float("nan")
    return metrics


def compute_confusion_matrix(y_true, y_pred) -> list[list[int]]:
    """Return confusion matrix as nested Python lists."""

    cm = confusion_matrix(y_true, y_pred)
    return [[int(v) for v in row] for row in cm.tolist()]
