requirements.txt
annotated-types==0.7.0
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
dask==2026.1.2
fonttools==4.61.1
fsspec==2026.2.0
greenlet==3.3.2
iniconfig==2.3.0
joblib==1.5.3
kiwisolver==1.4.9
locket==1.0.0
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
numpy==2.4.2
nvidia-nccl-cu12==2.29.3
packaging==26.0
pandas==3.0.1
partd==1.4.2
pillow==12.1.1
pluggy==1.6.0
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
SQLAlchemy==2.0.46
threadpoolctl==3.6.0
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
xgboost==3.2.0
